pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "django-cors-headers>=4.7.0",
]
